Shuffle structure, fix responses
